Add support for JDK versions beyond JDK8
Configuration which is specific to JDK versions higher than
JDK8 is handled in an extra profile (needed in particular to
enforce binary compatibility using the `release` flag which
is supported since JDK9).

Besides upgrading dependencies and plugins in the whole pom,
removed but still needed dependencies can be added in that
profile as well. This does not mean modularization works out
of the box.

The versions-plugin can now be used which makes version
discovery much more convenient. The older gmaven plugin has
been replaced by the newer groovy-maven-plugin.

The least required Maven version has been set to 3.5.0.